feat(modules): introduced other js module formats
