+++ .prettierignore
@@ -1,1 +1,1 @@
-*.min.js
+dist
