+++ demo/index.html
@@ -188,9 +188,11 @@
 			src="https://polyfill.io/v3/polyfill.min.js?flags=gated&features=IntersectionObserver%2C%2CmatchMedia"
 			crossorigin="anonymous"
 		></script>
+		<script type="module" src="index.js"></script>
 		<script
 			type="text/javascript"
-			src="../loading-attribute-polyfill.min.js"
+			src="../dist/loading-attribute-polyfill.umd.js"
+			nomodule
 		></script>
 	</body>
 </html>
